+++ brightwind/load/load.py
@@ -1450,7 +1450,52 @@ class LoadBrightHub:
             # Handle all request-related errors
             raise RuntimeError(f"An error occurred while fetching the cleaning log: {e}")
         return pd.read_csv(StringIO(cleaning_log_response.text))
-      
+
+    @staticmethod
+    def get_cleaning_rules(measurement_station_uuid):
+        """
+        Get the cleaning log from BrightHub for a particular measurement station.
+
+        :param measurement_station_uuid: A specific measurement station's uuid.
+        :type measurement_station_uuid:  str
+        :return:                         The cleaning logs for the measurement station.
+        :rtype:                          list(dict)
+
+        **Example usage**
+        ::
+            import brightwind as bw
+
+        To get the cleaning rules for a specific measurement station
+        ::
+            measurement_station_uuid='9344e576-6d5a-45f0-9750-2a7528ebfa14'
+            cleaning_rules_json = bw.LoadBrightHub.get_cleaning_rules(measurement_station_uuid)
+
+        Applying the cleaning to the timeseries data.
+        ::
+            data = bw.LoadBrightHub.get_data(measurement_station_uuid)
+
+            # Apply the cleaning rules to the data resulting in a dataset that is ready to work with.
+            data_clnd = bw.apply_cleaning_rules(data, cleaning_rules_json)
+
+        """
+        response = LoadBrightHub._brighthub_request(
+            url_end="/measurement-locations/{}/cleaning-rules".format(measurement_station_uuid))
+        response_json = response.json()
+
+        # error handling
+        if response.status_code == 400 and 'details' in response_json:  # request parameters invalid or missing
+            raise ValueError(f"{response_json.get('error', '')}. {response_json.get('details', '')}")
+        if response.status_code == 403 and 'details' in response_json:  # insufficient permissions or download limit
+            raise ValueError(f"{response_json.get('error', '')}. {response_json.get('details', '')}")
+        if response.status_code == 404 and 'details' in response_json:  # requested resource not found
+            raise ValueError(f"{response_json.get('error', '')}. {response_json.get('details', '')}")
+        if response.status_code == 500 and 'details' in response_json:  # unexpected server error
+            raise ValueError(f"{response_json.get('error', '')}. {response_json.get('details', '')}")
+        elif 'error' in response_json:
+            raise ValueError(f"Unexpected error: Status Code {response.status_code}. {response.text}")
+
+        return response_json
+
     @staticmethod
     def __get_reanalysis_nodes(reanalysis_name, min_latitude_ddeg, max_latitude_ddeg,
                                min_longitude_ddeg, max_longitude_ddeg):