+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Additional labels for pre-release and build metadata are available as extensions
 
 ## [2.X.0]
 1. Bug fix legend `plot_scatter_wspd` and `plot_scatter_wdir` functions and added tests (Issue [#443](https://github.com/brightwind-dev/brightwind/issues/443)).
-2. Write new note here (Issue [#4xx](https://github.com/brightwind-dev/brightwind/issues/4xx)).
+1. Bug fix calling `is_file` function (Issue [#447](https://github.com/brightwind-dev/brightwind/issues/447)).
 
 
 ## [2.2.1]